refactor: move around some configuration files
