+++ src/data_collection/disks/unix/macos/io_kit.rs
@@ -1,10 +1,8 @@
 mod bindings;
-
-mod io_iterator;
-pub use io_iterator::*;
-
-mod io_object;
-pub use io_object::*;
-
 mod io_disks;
+mod io_iterator;
+mod io_object;
+
 pub use io_disks::get_disks;
+pub use io_iterator::*;
+pub use io_object::*;
