+++ src/data_collection/disks/unix/macos/mod.rs
@@ -1,4 +1,4 @@
 mod counters;
-pub use counters::*;
-
 mod io_kit;
+
+pub use counters::*;
