+++ src/widgets.rs
@@ -1,20 +1,15 @@
-pub mod process_table;
-pub use process_table::*;
-
-pub mod temperature_table;
-pub use temperature_table::*;
-
-pub mod disk_table;
-pub use disk_table::*;
-
-pub mod cpu_graph;
-pub use cpu_graph::*;
-
-pub mod net_graph;
-pub use net_graph::*;
-
-pub mod mem_graph;
-pub use mem_graph::*;
-
 pub mod battery_widget;
+pub mod cpu_graph;
+pub mod disk_table;
+pub mod mem_graph;
+pub mod net_graph;
+pub mod process_table;
+pub mod temperature_table;
+
 pub use battery_widget::*;
+pub use cpu_graph::*;
+pub use disk_table::*;
+pub use mem_graph::*;
+pub use net_graph::*;
+pub use process_table::*;
+pub use temperature_table::*;
